requirements.txt
streamlit
matplotlib
seaborn
pandas
charset-normalizer
xlrd
openpyxl
pyperclip
pillow